requirements.txt
django==3.2.7
djangorestframework==3.12.4         #trabajar con rest
psycopg2==2.9.1                     #conector con postgrade
gunicorn==20.1.0                    #correr el servidor en produccion
django-heroku==0.3.1                #desplegar django en heroku
django-rest-auth==0.9.5
django-allauth
django-rest-auth[with_social]